# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation for my Laravel project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employes": {
            "get": {
                "tags": [
                    "Employes"
                ],
                "summary": "Get all employees",
                "description": "Returns a list of all employees",
                "operationId": "7f4ab6a26380d98780b1b17ef2e473ae",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employe"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No employees found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Employes"
                ],
                "summary": "Create a new employee",
                "description": "Add a new employee to the system",
                "operationId": "191907e0e0347b310ca8c9ceccb28ad8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_utilisateur",
                                    "poste",
                                    "date_embauche",
                                    "salaire",
                                    "est_actif"
                                ],
                                "properties": {
                                    "id_utilisateur": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "poste": {
                                        "type": "string",
                                        "example": "Développeur Backend"
                                    },
                                    "date_embauche": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-20"
                                    },
                                    "salaire": {
                                        "type": "number",
                                        "example": 1200000
                                    },
                                    "est_actif": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employe"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/employes/{id}": {
            "get": {
                "tags": [
                    "Employes"
                ],
                "summary": "Get a single employee by ID",
                "operationId": "e0e6bace9c7b75a759f0aa559bbc3d3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employe"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employes"
                ],
                "summary": "Update an employee",
                "operationId": "6790ebda216c4018c834522a0a272be6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "poste": {
                                        "type": "string",
                                        "example": "Développeur Senior"
                                    },
                                    "date_embauche": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-21"
                                    },
                                    "salaire": {
                                        "type": "number",
                                        "example": 1500000
                                    },
                                    "est_actif": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employe"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employes"
                ],
                "summary": "Delete an employee",
                "operationId": "39846ba0f029b3786eda0683c811e1a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee deleted successfully"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Employe": {
                "title": "Employe",
                "required": [
                    "id_utilisateur",
                    "poste",
                    "date_embauche",
                    "salaire",
                    "est_actif"
                ],
                "properties": {
                    "id_employe": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_utilisateur": {
                        "type": "integer",
                        "example": 1
                    },
                    "poste": {
                        "type": "string",
                        "example": "Développeur Backend"
                    },
                    "date_embauche": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-20"
                    },
                    "salaire": {
                        "type": "number",
                        "example": 1200000
                    },
                    "est_actif": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}